server/connecthand.cc

+++ server/connecthand.cc
@@ -848,7 +848,7 @@ bool attach_connection_to_player(connection_t *pconn,
 {
   /* if pplayer is NULL, attach to first non-connected player slot */
   if (!pplayer) {
-    if (game.info.nplayers >= game.info.max_players
+    if ((int)game.info.nplayers >= game.info.max_players
         || game.info.nplayers >= MAX_NUM_PLAYERS + MAX_NUM_BARBARIANS
         || !can_control_a_player(pconn, TRUE)) {
       return FALSE;
